services/backend/requirements.txt
python-multipart
torch==2.0.1
fastapi
Pillow
numpy
uvicorn
gradio